Hopefully fix setting chart y-range.
Previously, using `std::to_string(double)` was causing small number like 1.0E-12 to print to strings like "0.0000", which messed up the view.
Also logic for overriding user directed y-range was a bit suspect - although its still not likely totally ironed out.

+++ InterSpec/InterSpec.h
@@ -593,7 +593,7 @@ public:
    Returns true if successful, or false if the request couldnt be fully honored
    (e.g., a negative lower counts was specified, but its currently log-scale)
    */
-  bool setYAxisRange( float lower_counts, float upper_counts );
+  bool setYAxisRange( double lower_counts, double upper_counts );
 
   /** When displaying the spectrum in log-y, sets the lower y-axis value to show, if there are channels with zero counts. */
   bool setLogYAxisMin( const double lower_value );
